Naming and formatting changes. Removed the fluid storage housing. Implemented #3434 for FluidSorters
