+++ src/api/java/appeng/api/definitions/IMaterials.java
@@ -142,6 +142,4 @@ public interface IMaterials
 	IItemDefinition fluidCell16kPart();
 
 	IItemDefinition fluidCell64kPart();
-
-	IItemDefinition emptyFluidStorageCell();
 }
