+++ src/main/java/appeng/core/api/definitions/ApiParts.java
@@ -77,16 +77,17 @@ public final class ApiParts implements IParts
 	private final IItemDefinition terminal;
 	private final IItemDefinition storageMonitor;
 	private final IItemDefinition conversionMonitor;
-	private final IItemDefinition importBusFluid;
-	private final IItemDefinition exportBusFluid;
+	private final IItemDefinition fluidImportBus;
+	private final IItemDefinition fluidExportBus;
 	private final IItemDefinition fluidTerminal;
-	private final IItemDefinition storageBusFluid;
-	private final IItemDefinition levelEmitterFluid;
+	private final IItemDefinition fluidStorageBus;
 
 	public ApiParts( FeatureFactory registry, PartModels partModels )
 	{
 		final ItemPart itemPart = new ItemPart();
-		registry.item( "part", () -> itemPart ).rendering( new ItemPartRendering( partModels, itemPart ) ).build();
+		registry.item( "part", () -> itemPart )
+				.rendering( new ItemPartRendering( partModels, itemPart ) )
+				.build();
 
 		// Register all part models
 		for( PartType partType : PartType.values() )
@@ -112,7 +113,8 @@ public final class ApiParts implements IParts
 		this.iface = new DamagedItemDefinition( "part.interface", itemPart.createPart( PartType.INTERFACE ) );
 		this.levelEmitter = new DamagedItemDefinition( "part.level_emitter", itemPart.createPart( PartType.LEVEL_EMITTER ) );
 		this.annihilationPlane = new DamagedItemDefinition( "part.plane.annihilation", itemPart.createPart( PartType.ANNIHILATION_PLANE ) );
-		this.identityAnnihilationPlane = new DamagedItemDefinition( "part.plane.annihiliation.identity", itemPart.createPart( PartType.IDENTITY_ANNIHILATION_PLANE ) );
+		this.identityAnnihilationPlane = new DamagedItemDefinition( "part.plane.annihiliation.identity",
+				itemPart.createPart( PartType.IDENTITY_ANNIHILATION_PLANE ) );
 		this.formationPlane = new DamagedItemDefinition( "part.plane.formation", itemPart.createPart( PartType.FORMATION_PLANE ) );
 		this.p2PTunnelME = new DamagedItemDefinition( "part.tunnel.me", itemPart.createPart( PartType.P2P_TUNNEL_ME ) );
 		this.p2PTunnelRedstone = new DamagedItemDefinition( "part.tunnel.redstone", itemPart.createPart( PartType.P2P_TUNNEL_REDSTONE ) );
@@ -133,11 +135,10 @@ public final class ApiParts implements IParts
 		this.terminal = new DamagedItemDefinition( "part.terminal", itemPart.createPart( PartType.TERMINAL ) );
 		this.storageMonitor = new DamagedItemDefinition( "part.monitor.storage", itemPart.createPart( PartType.STORAGE_MONITOR ) );
 		this.conversionMonitor = new DamagedItemDefinition( "part.monitor.conversion", itemPart.createPart( PartType.CONVERSION_MONITOR ) );
-		this.importBusFluid = new DamagedItemDefinition( "part.bus.import.fluid", itemPart.createPart( PartType.IMPORT_BUS_FLUID ) );
-		this.exportBusFluid = new DamagedItemDefinition( "part.bus.export.fluid", itemPart.createPart( PartType.EXPORT_BUS_FLUID ) );
+		this.fluidImportBus = new DamagedItemDefinition( "part.bus.import.fluid", itemPart.createPart( PartType.FLUID_IMPORT_BUS ) );
+		this.fluidExportBus = new DamagedItemDefinition( "part.bus.export.fluid", itemPart.createPart( PartType.FLUID_EXPORT_BUS ) );
 		this.fluidTerminal = new DamagedItemDefinition( "part.terminal.fluid", itemPart.createPart( PartType.FLUID_TERMINAL ) );
-		this.storageBusFluid = new DamagedItemDefinition( "part.bus.storage.fluid", itemPart.createPart( PartType.STORAGE_BUS_FLUID ) );
-		this.levelEmitterFluid = new DamagedItemDefinition( "part.level_emitter.fluid", itemPart.createPart( PartType.LEVEL_EMITTER_FLUID ) );
+		this.fluidStorageBus = new DamagedItemDefinition( "part.bus.storage.fluid", itemPart.createPart( PartType.FLUID_STORAGE_BUS ) );
 	}
 
 	private static AEColoredItemDefinition constructColoredDefinition( final ItemPart target, final PartType type )
@@ -395,26 +396,20 @@ public final class ApiParts implements IParts
 	}
 
 	@Override
-	public IItemDefinition importBusFluids()
+	public IItemDefinition fluidImportBus()
 	{
-		return this.importBusFluid;
+		return this.fluidImportBus;
 	}
 
 	@Override
-	public IItemDefinition exportBusFluids()
+	public IItemDefinition fluidExportBus()
 	{
-		return this.exportBusFluid;
+		return this.fluidExportBus;
 	}
 
 	@Override
-	public IItemDefinition storageBusFluids()
+	public IItemDefinition fluidStorageBus()
 	{
-		return this.storageBusFluid;
-	}
-
-	@Override
-	public IItemDefinition levelEmitterFluids()
-	{
-		return this.levelEmitterFluid;
+		return this.fluidStorageBus;
 	}
 }
