+++ src/main/java/appeng/core/features/AEFeature.java
@@ -76,7 +76,6 @@ public enum AEFeature
 
 	INTERFACE( "Interface", Constants.CATEGORY_NETWORK_BUSES ),
 	LEVEL_EMITTER( "LevelEmitter", Constants.CATEGORY_NETWORK_BUSES ),
-	LEVEL_EMITTER_FLUID( "LevelEmitterFluid", Constants.CATEGORY_NETWORK_BUSES ),
 	FLUID_TERMINAL( "FluidTerminal", Constants.CATEGORY_NETWORK_BUSES ),
 	CRAFTING_TERMINAL( "CraftingTerminal", Constants.CATEGORY_NETWORK_BUSES ),
 	TERMINAL( "Terminal", Constants.CATEGORY_NETWORK_BUSES ),
@@ -86,11 +85,11 @@ public enum AEFeature
 	ANNIHILATION_PLANE( "AnnihilationPlane", Constants.CATEGORY_NETWORK_BUSES ),
 	IDENTITY_ANNIHILATION_PLANE( "IdentityAnnihilationPlane", Constants.CATEGORY_NETWORK_BUSES ),
 	IMPORT_BUS( "ImportBus", Constants.CATEGORY_NETWORK_BUSES ),
-	IMPORT_BUS_FLUID( "ImportBusFluid", Constants.CATEGORY_NETWORK_BUSES ),
+	FLUID_IMPORT_BUS( "FluidImportBus", Constants.CATEGORY_NETWORK_BUSES ),
 	EXPORT_BUS( "ExportBus", Constants.CATEGORY_NETWORK_BUSES ),
-	EXPORT_BUS_FLUID( "ExportBusFluid", Constants.CATEGORY_NETWORK_BUSES ),
+	FLUID_EXPORT_BUS( "FluidExportBus", Constants.CATEGORY_NETWORK_BUSES ),
 	STORAGE_BUS( "StorageBus", Constants.CATEGORY_NETWORK_BUSES ),
-	STORAGE_BUS_FLUID( "StorageBusFluid", Constants.CATEGORY_NETWORK_BUSES ),
+	FLUID_STORAGE_BUS( "FluidStorageBus", Constants.CATEGORY_NETWORK_BUSES ),
 	PART_CONVERSION_MONITOR( "PartConversionMonitor", Constants.CATEGORY_NETWORK_BUSES ),
 	TOGGLE_BUS( "ToggleBus", Constants.CATEGORY_NETWORK_BUSES ),
 	PANELS( "Panels", Constants.CATEGORY_NETWORK_BUSES ),
