+++ src/main/java/appeng/util/FluidSorters.java
@@ -84,50 +84,12 @@ public class FluidSorters
 		{
 			if( getDirection() == SortDir.ASCENDING )
 			{
-				return compareLong( o2.getStackSize(), o1.getStackSize() );
+				return Long.compare( o2.getStackSize(), o1.getStackSize() );
 			}
-			return compareLong( o1.getStackSize(), o2.getStackSize() );
+			return Long.compare( o1.getStackSize(), o2.getStackSize() );
 		}
 	};
 
-	public static int compareInt( final int a, final int b )
-	{
-		if( a == b )
-		{
-			return 0;
-		}
-		if( a < b )
-		{
-			return -1;
-		}
-		return 1;
-	}
-
-	public static int compareLong( final long a, final long b )
-	{
-		if( a == b )
-		{
-			return 0;
-		}
-		if( a < b )
-		{
-			return -1;
-		}
-		return 1;
-	}
-
-	public static int compareDouble( final double a, final double b )
-	{
-		if( a == b )
-		{
-			return 0;
-		}
-		if( a < b )
-		{
-			return -1;
-		}
-		return 1;
-	}
 
 	private static SortDir getDirection()
 	{
