+++ src/main/resources/assets/appliedenergistics2/lang/en_us.lang
@@ -490,17 +490,17 @@ item.appliedenergistics2.multi_part.semi_dark_monitor.name=Illuminated Panel
 item.appliedenergistics2.multi_part.monitor.name=Bright Illuminated Panel
 item.appliedenergistics2.multi_part.dark_monitor.name=Dark Illuminated Panel
 item.appliedenergistics2.multi_part.export_bus.name=ME Export Bus
-item.appliedenergistics2.multi_part.export_bus_fluid.name=ME Fluid Export Bus
+item.appliedenergistics2.multi_part.fluid_export_bus.name=ME Fluid Export Bus
 item.appliedenergistics2.multi_part.formation_plane.name=ME Formation Plane
 item.appliedenergistics2.multi_part.import_bus.name=ME Import Bus
-item.appliedenergistics2.multi_part.import_bus_fluid.name=ME Fluid Import Bus
+item.appliedenergistics2.multi_part.fluid_import_bus.name=ME Fluid Import Bus
 item.appliedenergistics2.multi_part.interface.name=ME Interface
 item.appliedenergistics2.multi_part.level_emitter.name=ME Level Emitter
 item.appliedenergistics2.multi_part.p2p_tunnel.name=P2P Tunnel
 item.appliedenergistics2.multi_part.pattern_terminal.name=ME Pattern Terminal
 item.appliedenergistics2.multi_part.quartz_fiber.name=Quartz Fiber
 item.appliedenergistics2.multi_part.storage_bus.name=ME Storage Bus
-item.appliedenergistics2.multi_part.storage_bus_fluid.name=ME Fluid Storage Bus
+item.appliedenergistics2.multi_part.fluid_storage_bus.name=ME Fluid Storage Bus
 item.appliedenergistics2.multi_part.terminal.name=ME Terminal
 item.appliedenergistics2.multi_part.inverted_toggle_bus.name=ME Inverted Toggle Bus
 item.appliedenergistics2.multi_part.toggle_bus.name=ME Toggle Bus
