--- src/main/resources/assets/appliedenergistics2/models/item/material_empty_fluid_storage_cell.json
@@ -1,6 +1,0 @@
-{
-    "parent": "item/generated",
-    "textures": {
-        "layer0": "appliedenergistics2:items/material_empty_fluid_storage_cell"
-    }
-}
