--- src/main/resources/assets/appliedenergistics2/models/item/part/export_bus_fluid.json
@@ -1,106 +1,0 @@
-{
-    "parent": "appliedenergistics2:item/part/part",
-    "textures": {
-        "sides": "appliedenergistics2:parts/export_bus_sides",
-        "back": "appliedenergistics2:parts/monitor_back",
-        "front": "appliedenergistics2:items/part/export_bus"
-    },
-    "elements": [
-        {
-            "from": [
-                4,
-                4,
-                8
-            ],
-            "to": [
-                12,
-                12,
-                10
-            ],
-            "faces": {
-                "north": {
-                    "texture": "#front"
-                },
-                "south": {
-                    "texture": "#back"
-                },
-                "east": {
-                    "texture": "#sides"
-                },
-                "west": {
-                    "texture": "#sides"
-                },
-                "up": {
-                    "texture": "#sides"
-                },
-                "down": {
-                    "texture": "#sides"
-                }
-            }
-        },
-        {
-            "from": [
-                5,
-                5,
-                7
-            ],
-            "to": [
-                11,
-                11,
-                8
-            ],
-            "faces": {
-                "north": {
-                    "texture": "#front"
-                },
-                "south": {
-                    "texture": "#back"
-                },
-                "east": {
-                    "texture": "#sides"
-                },
-                "west": {
-                    "texture": "#sides"
-                },
-                "up": {
-                    "texture": "#sides"
-                },
-                "down": {
-                    "texture": "#sides"
-                }
-            }
-        },
-        {
-            "from": [
-                6,
-                6,
-                6
-            ],
-            "to": [
-                10,
-                10,
-                7
-            ],
-            "faces": {
-                "north": {
-                    "texture": "#front"
-                },
-                "south": {
-                    "texture": "#back"
-                },
-                "east": {
-                    "texture": "#sides"
-                },
-                "west": {
-                    "texture": "#sides"
-                },
-                "up": {
-                    "texture": "#sides"
-                },
-                "down": {
-                    "texture": "#sides"
-                }
-            }
-        }
-    ]
-}
