+++ src/main/resources/assets/appliedenergistics2/models/item/part/fluid_import_bus.json
@@ -1,0 +1,106 @@
+{
+    "parent": "appliedenergistics2:item/part/part",
+    "textures": {
+        "sides": "appliedenergistics2:parts/import_bus_sides",
+        "back": "appliedenergistics2:parts/monitor_back",
+        "front": "appliedenergistics2:items/part/import_bus"
+    },
+    "elements": [
+        {
+            "from": [
+                3,
+                3,
+                8
+            ],
+            "to": [
+                13,
+                13,
+                9
+            ],
+            "faces": {
+                "north": {
+                    "texture": "#front"
+                },
+                "south": {
+                    "texture": "#back"
+                },
+                "east": {
+                    "texture": "#sides"
+                },
+                "west": {
+                    "texture": "#sides"
+                },
+                "up": {
+                    "texture": "#sides"
+                },
+                "down": {
+                    "texture": "#sides"
+                }
+            }
+        },
+        {
+            "from": [
+                4,
+                4,
+                9
+            ],
+            "to": [
+                12,
+                12,
+                10
+            ],
+            "faces": {
+                "north": {
+                    "texture": "#front"
+                },
+                "south": {
+                    "texture": "#back"
+                },
+                "east": {
+                    "texture": "#sides"
+                },
+                "west": {
+                    "texture": "#sides"
+                },
+                "up": {
+                    "texture": "#sides"
+                },
+                "down": {
+                    "texture": "#sides"
+                }
+            }
+        },
+        {
+            "from": [
+                5,
+                5,
+                10
+            ],
+            "to": [
+                11,
+                11,
+                11
+            ],
+            "faces": {
+                "north": {
+                    "texture": "#front"
+                },
+                "south": {
+                    "texture": "#back"
+                },
+                "east": {
+                    "texture": "#sides"
+                },
+                "west": {
+                    "texture": "#sides"
+                },
+                "up": {
+                    "texture": "#sides"
+                },
+                "down": {
+                    "texture": "#sides"
+                }
+            }
+        }
+    ]
+}
