--- src/main/resources/assets/appliedenergistics2/models/item/part/storage_bus_fluid.json
@@ -1,106 +1,0 @@
-{
-    "parent": "appliedenergistics2:item/part/part",
-    "textures": {
-        "sides": "appliedenergistics2:parts/storage_bus_sides",
-        "back": "appliedenergistics2:parts/storage_bus_back",
-        "front": "appliedenergistics2:items/part/storage_bus"
-    },
-    "elements": [
-        {
-            "from": [
-                3,
-                3,
-                6
-            ],
-            "to": [
-                13,
-                13,
-                7
-            ],
-            "faces": {
-                "north": {
-                    "texture": "#front"
-                },
-                "south": {
-                    "texture": "#back"
-                },
-                "east": {
-                    "texture": "#sides"
-                },
-                "west": {
-                    "texture": "#sides"
-                },
-                "up": {
-                    "texture": "#sides"
-                },
-                "down": {
-                    "texture": "#sides"
-                }
-            }
-        },
-        {
-            "from": [
-                2,
-                2,
-                7
-            ],
-            "to": [
-                14,
-                14,
-                8
-            ],
-            "faces": {
-                "north": {
-                    "texture": "#front"
-                },
-                "south": {
-                    "texture": "#back"
-                },
-                "east": {
-                    "texture": "#sides"
-                },
-                "west": {
-                    "texture": "#sides"
-                },
-                "up": {
-                    "texture": "#sides"
-                },
-                "down": {
-                    "texture": "#sides"
-                }
-            }
-        },
-        {
-            "from": [
-                5,
-                5,
-                8
-            ],
-            "to": [
-                11,
-                11,
-                10
-            ],
-            "faces": {
-                "north": {
-                    "texture": "#front"
-                },
-                "south": {
-                    "texture": "#back"
-                },
-                "east": {
-                    "texture": "#sides"
-                },
-                "west": {
-                    "texture": "#sides"
-                },
-                "up": {
-                    "texture": "#sides"
-                },
-                "down": {
-                    "texture": "#sides"
-                }
-            }
-        }
-    ]
-}
