+++ src/main/resources/assets/appliedenergistics2/models/part/fluid_export_bus_has_channel.json
