+++ src/main/resources/assets/appliedenergistics2/models/part/fluid_export_bus_on.json
