+++ src/main/resources/assets/appliedenergistics2/models/part/fluid_storage_bus_off.json
